[GREEN] move process data logic of each chart to each chart file
